Add message template for state
Untill now 'payload' was abused to put the whole message in. Now a clean message
template is seperated from the real payload, which is basically 'turnOn/Off'.
This makes it possible now to reflect the state in home-assistent (although this
only works for on/off, as brightness changes are not published on the sphere's mqtt
for some reason).